--- api/src/TestEngineSettings.cs
@@ -1,30 +1,0 @@
-﻿namespace GdUnit4;
-
-/// <summary>
-///     Configuration settings for the GdUnit4 test engine.
-///     Contains runtime parameters and execution options that control test behavior.
-/// </summary>
-public sealed class TestEngineSettings
-{
-    /// <summary>
-    ///     Additional Godot runtime parameters. These are passed to the Godot executable when running tests.
-    /// </summary>
-    public string? Parameters { get; init; }
-
-
-    /// <summary>
-    ///     When set to true, standard output (stdout) from test cases is captured and included in the test result. This can be
-    ///     useful for debugging.
-    ///     Default: false
-    /// </summary>
-    public bool CaptureStdOut { get; init; }
-
-    public int MaxCpuCount { get; init; }
-
-
-    /// <summary>
-    ///     The test session timout in milliseconds.
-    ///     Default: 30s
-    /// </summary>
-    public int SessionTimeout { get; init; } = 30000;
-}
