+++ api/src/core/runners/InOutPipeProxy.cs
@@ -9,6 +9,8 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
+using Api;
+
 using Commands;
 
 using Newtonsoft.Json;
