+++ test/gdunit4_testadapter/GdUnit4TestRunnerScene.cs
@@ -6,6 +6,8 @@ namespace GdUnit4.TestRunner;
 
 using System;
 
+using Api;
+
 using Core.Runners;
 
 using Godot;
